offboad_py is inline to the offboard_cpp
