--- src/examples/offboard_py/package.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>offboard</name>
-  <version>0.0.0</version>
-  <description>offboard controller with ros2</description>
-  <maintainer email="harisankar995@gmail.com">hari</maintainer>
-  <license>TODO: License declaration</license>
-  
-  <exec_depend>rclpy</exec_depend>
-  <exec_depend>px4_msgs</exec_depend>
-  <exec_depend>rosidl_default_runtime</exec_depend>
-  <exec_depend>px4_interfaces</exec_depend>
-  
-  <depend>builtin_interfaces</depend>
-  <depend>geometry_msgs</depend>
-  <depend>sensor_msgs</depend>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
