--- src/examples/offboard_py/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/offboard_py
-[install]
-install-scripts=$base/lib/offboard_py
