--- src/examples/offboard_py/setup.py
@@ -1,26 +1,0 @@
-from setuptools import setup
-
-package_name = 'offboard_py'
-
-setup(
-    name=package_name,
-    version='0.0.1',
-    packages=[package_name],
-    data_files=[
-        ('share/ament_index/resource_index/packages',
-            ['resource/' + package_name]),
-        ('share/' + package_name, ['package.xml']),
-    ],
-    install_requires=['setuptools'],
-    zip_safe=True,
-    maintainer='hari',
-    maintainer_email='harisankar995@gmail.com',
-    description='offboard controller with ros2',
-    license='TODO: License declaration',
-    tests_require=['pytest'],
-    entry_points={
-        'console_scripts': [
-            'offboard = offboard_py.offboard:main'
-        ],
-    },
-)
