--- src/examples/px4_interfaces/msg/Trajectory.msg
@@ -1,4 +1,0 @@
-float32 x
-float32 y
-float32 z
-float32 yaw
