--- src/examples/px4_interfaces/package.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>px4_interfaces</name>
-  <version>0.0.0</version>
-  <description>Interfaces for PX4</description>
-  <maintainer email="harisankar995@gmail.com">hari</maintainer>
-  <license>Apache License 2.0</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>  
-  <build_depend>rosidl_default_generators</build_depend>
-  <build_depend>builtin_interfaces</build_depend>
-
-  <depend>ros_environment</depend>
-
-  <exec_depend>rosidl_default_runtime</exec_depend>
-  <exec_depend>builtin_interfaces</exec_depend>
-
-  <test_depend>ament_lint_common</test_depend>
-
-  <member_of_group>rosidl_interface_packages</member_of_group>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
