--- src/examples/py_publisher/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/py_publisher
-[install]
-install-scripts=$base/lib/py_publisher
